use provider to manage web3 connection and activity state
